pyproject.toml
[tool.poetry]
name = "lampsz-app"
version = "0.1.0"
description = ""
authors = ["Le Rong Shi <shilerong@gmail.com>"]
readme = "README.md"
packages = [{include = "lampsz_app"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
flake8 = "^5.0.4"
mypy = "^0.982"
flake8-isort = "^5.0.0"
black = "^22.10.0"
pre-commit = "^2.20.0"
psycopg2 = "^2.9.5"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"
sentry-sdk = "^1.10.1"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
google-api-python-client = "^2.66.0"
google-auth = "^2.14.1"
google-auth-oauthlib = "^0.7.1"
google-auth-httplib2 = "^0.1.0"
coverage = "^6.5.0"
djangorestframework-camel-case = "^1.3.0"
pillow = "^9.3.0"
spacy = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
